contrib/IBM/sarama.v1: add WrapConsumerGroupHandler to trace consumer groups
